Merged PR 232: Add Winkyface to Top Mink
Turns out it's already implemented. But CSS is overriding colour to white.

It turns out Icon has a getColor which it was getting from its parent wrapper. But the * white was overriding it from a parent.

This means some icons changed colour. I checked the different icons and found Flag changed colour because its colour was set to 'secondary' which is meant to be gray2